+++ packages/cli/src/commands/sandbox/sandbox-secret/sandbox_secret_set_command.test.ts
@@ -6,6 +6,8 @@ import assert from 'node:assert';
 import { SandboxBackendIdResolver } from '../sandbox_id_resolver.js';
 import { SecretIdentifier, getSecretClient } from '@aws-amplify/backend-secret';
 import { SandboxSecretSetCommand } from './sandbox_secret_set_command.js';
+import { ReadStream } from 'node:tty';
+import { PassThrough } from 'node:stream';
 
 const testSecretName = 'testSecretName';
 const testSecretValue = 'testSecretValue';
@@ -34,9 +36,13 @@ void describe('sandbox secret set command', () => {
       }),
   } as SandboxBackendIdResolver;
 
+  const mockReadStream = new PassThrough();
+  (mockReadStream as unknown as ReadStream).isTTY = true; // Fake a TTY in tests
+
   const sandboxSecretSetCmd = new SandboxSecretSetCommand(
     sandboxIdResolver,
-    secretClient
+    secretClient,
+    mockReadStream as unknown as ReadStream
   );
 
   const parser = yargs().command(
@@ -89,6 +95,41 @@ void describe('sandbox secret set command', () => {
         type: 'sandbox',
         namespace: testBackendId,
         name: 'anotherName',
+      },
+      testSecretName,
+      testSecretValue,
+    ]);
+  });
+
+  void it('sets a secret using redirection', async () => {
+    const readStream = new PassThrough(); // 'isTTY' flag doesn't exist on PassThrough stream.
+
+    const sandboxSecretSetCmdWithStream = new SandboxSecretSetCommand(
+      sandboxIdResolver,
+      secretClient,
+      readStream as unknown as ReadStream
+    );
+
+    const parserWithStream = yargs().command(
+      sandboxSecretSetCmdWithStream as unknown as CommandModule
+    );
+
+    const commandRunnerWithStream = new TestCommandRunner(parserWithStream);
+
+    const setCommandPromise = commandRunnerWithStream.runCommand(
+      `set ${testSecretName}`
+    );
+    readStream.write(testSecretValue);
+    readStream.end();
+    await setCommandPromise;
+
+    assert.equal(secretSetMock.mock.callCount(), 1);
+
+    assert.deepStrictEqual(secretSetMock.mock.calls[0].arguments, [
+      {
+        type: 'sandbox',
+        namespace: testBackendId,
+        name: testSandboxName,
       },
       testSecretName,
       testSecretValue,