+++ packages/client-config/src/client-config-writer/client_config_to_mobile_legacy_converter.test.ts
@@ -37,6 +37,15 @@ void describe('client config converter', () => {
       ],
       aws_cognito_mfa_configuration: 'test_mfa_configuration',
       aws_cognito_mfa_types: ['test_mfa_type_1', 'test_mfa_type_2'],
+      aws_cognito_social_providers: ['GOOGLE', 'FACEBOOK', 'AMAZON', 'APPLE'],
+
+      oauth: {
+        clientId: '',
+        domain: 'test_domain',
+        scope: ['test_scope_1', 'test_scope_2'],
+        redirectSignIn: 'test_redirect_sign_in',
+        redirectSignOut: 'test_redirect_sign_out',
+      },
     };
     const expectedMobileConfig: ClientConfigMobile = {
       UserAgent: expectedUserAgent,
@@ -70,6 +79,7 @@ void describe('client config converter', () => {
                   'test_signup_attribute_1',
                   'test_signup_attribute_2',
                 ],
+                socialProviders: ['GOOGLE', 'FACEBOOK', 'AMAZON', 'APPLE'],
                 usernameAttributes: [
                   'test_username_attribute_1',
                   'test_username_attribute_2',
@@ -82,6 +92,13 @@ void describe('client config converter', () => {
                   'test_verification_mechanism_1',
                   'test_verification_mechanism_2',
                 ],
+                OAuth: {
+                  WebDomain: 'test_domain',
+                  AppClientId: 'test_user_pool_app_client_id',
+                  Scopes: ['test_scope_1', 'test_scope_2'],
+                  SignInRedirectURI: 'test_redirect_sign_in',
+                  SignOutRedirectURI: 'test_redirect_sign_out',
+                },
               },
             },
           },
@@ -136,6 +153,7 @@ void describe('client config converter', () => {
       aws_appsync_authenticationType: 'API_KEY',
       aws_appsync_additionalAuthenticationTypes:
         'AMAZON_COGNITO_USER_POOLS,AWS_IAM',
+      aws_cognito_social_providers: ['GOOGLE', 'FACEBOOK', 'AMAZON', 'APPLE'],
     };
     const expectedMobileConfig: ClientConfigMobile = {
       UserAgent: expectedUserAgent,
@@ -166,6 +184,7 @@ void describe('client config converter', () => {
                 mfaConfiguration: undefined,
                 mfaTypes: undefined,
                 signupAttributes: [],
+                socialProviders: ['GOOGLE', 'FACEBOOK', 'AMAZON', 'APPLE'],
                 usernameAttributes: [],
                 passwordProtectionSettings: {
                   passwordPolicyCharacters: [],
@@ -185,14 +204,14 @@ void describe('client config converter', () => {
               data_AWS_IAM: {
                 ApiUrl: 'https://test_api_endpoint.amazon.com',
                 Region: 'test_app_sync_region',
-                AuthMode: 'API_KEY',
+                AuthMode: 'AWS_IAM',
                 ApiKey: 'test_api_key',
                 ClientDatabasePrefix: 'data_AWS_IAM',
               },
               data_AMAZON_COGNITO_USER_POOLS: {
                 ApiUrl: 'https://test_api_endpoint.amazon.com',
                 Region: 'test_app_sync_region',
-                AuthMode: 'API_KEY',
+                AuthMode: 'AMAZON_COGNITO_USER_POOLS',
                 ApiKey: 'test_api_key',
                 ClientDatabasePrefix: 'data_AMAZON_COGNITO_USER_POOLS',
               },