+++ packages/client-config/src/client-config-writer/client_config_to_mobile_legacy_converter.ts
@@ -69,11 +69,24 @@ export class ClientConfigMobileConverter {
                   clientConfig.aws_cognito_username_attributes ?? [],
                 verificationMechanisms:
                   clientConfig.aws_cognito_verification_mechanisms ?? [],
+                socialProviders:
+                  clientConfig.aws_cognito_social_providers ?? [],
               },
             },
           },
         },
       };
+      if (clientConfig.oauth) {
+        authConfig.plugins.awsCognitoAuthPlugin.Auth.Default.OAuth = {
+          WebDomain: clientConfig.oauth.domain,
+          Scopes: clientConfig.oauth.scope,
+          SignInRedirectURI: clientConfig.oauth.redirectSignIn,
+          SignOutRedirectURI: clientConfig.oauth.redirectSignOut,
+          AppClientId:
+            clientConfig.oauth.clientId ||
+            clientConfig.aws_user_pools_web_client_id,
+        };
+      }
       mobileConfig.auth = authConfig;
     }
 
@@ -116,7 +129,7 @@ export class ClientConfigMobileConverter {
             ] = {
               ApiUrl: clientConfig.aws_appsync_graphqlEndpoint,
               Region: clientConfig.aws_appsync_region,
-              AuthMode: clientConfig.aws_appsync_authenticationType,
+              AuthMode: additionalAuthenticationType,
               ApiKey: clientConfig.aws_appsync_apiKey,
               ClientDatabasePrefix: `data_${additionalAuthenticationType}`,
             };