+++ packages/backend/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @aws-amplify/backend
 
+## 1.0.3
+
+### Patch Changes
+
+- 530bf2c: chore: disambiguate url imports and import from node:url explicitly
+- Updated dependencies [1146bbd]
+- Updated dependencies [530bf2c]
+- Updated dependencies [925f97d]
+  - @aws-amplify/client-config@1.0.3
+  - @aws-amplify/backend-function@1.0.3
+  - @aws-amplify/backend-storage@1.0.2
+  - @aws-amplify/platform-core@1.0.1
+  - @aws-amplify/backend-data@1.0.2
+
 ## 1.0.2
 
 ### Patch Changes